// frontend/src/api/generated/models/picking-instruction-summary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PickingInstructionStatus } from './picking-instruction-status';

/**
 * 
 * @export
 * @interface PickingInstructionSummary
 */
export interface PickingInstructionSummary {
    /**
     * ピッキング指示ID
     * @type {number}
     * @memberof PickingInstructionSummary
     */
    'id'?: number;
    /**
     * 指示番号
     * @type {string}
     * @memberof PickingInstructionSummary
     */
    'instructionNumber'?: string;
    /**
     * 倉庫ID
     * @type {number}
     * @memberof PickingInstructionSummary
     */
    'warehouseId'?: number;
    /**
     * 倉庫名
     * @type {string}
     * @memberof PickingInstructionSummary
     */
    'warehouseName'?: string;
    /**
     * エリアID
     * @type {number}
     * @memberof PickingInstructionSummary
     */
    'areaId'?: number | null;
    /**
     * エリア名
     * @type {string}
     * @memberof PickingInstructionSummary
     */
    'areaName'?: string | null;
    /**
     * 
     * @type {PickingInstructionStatus}
     * @memberof PickingInstructionSummary
     */
    'status'?: PickingInstructionStatus;
    /**
     * 明細件数
     * @type {number}
     * @memberof PickingInstructionSummary
     */
    'lineCount'?: number;
    /**
     * 作成日時
     * @type {string}
     * @memberof PickingInstructionSummary
     */
    'createdAt'?: string;
    /**
     * 作成者名
     * @type {string}
     * @memberof PickingInstructionSummary
     */
    'createdByName'?: string;
}