// frontend/src/api/generated/models/picking-instruction-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { PickingInstructionSummary } from './picking-instruction-summary';

/**
 * 
 * @export
 * @interface PickingInstructionPageResponse
 */
export interface PickingInstructionPageResponse {
    /**
     * 
     * @type {Array<PickingInstructionSummary>}
     * @memberof PickingInstructionPageResponse
     */
    'content'?: Array<PickingInstructionSummary>;
    /**
     * 
     * @type {number}
     * @memberof PickingInstructionPageResponse
     */
    'page'?: number;
    /**
     * 
     * @type {number}
     * @memberof PickingInstructionPageResponse
     */
    'size'?: number;
    /**
     * 
     * @type {number}
     * @memberof PickingInstructionPageResponse
     */
    'totalElements'?: number;
    /**
     * 
     * @type {number}
     * @memberof PickingInstructionPageResponse
     */
    'totalPages'?: number;
}